move old image upload view functions to miscellaneous folder
