+++ backend/api/views/__init__.py
@@ -1,5 +1,7 @@
 from .factories_cr import get_nearby_or_create_factories
 from .factories_u import update_factory_attribute
-from .image_c import post_image
-from .factory_image_c import post_factory_image
 from .factory_report_record_r import get_factory_report
+
+# To be deprecated
+from .miscellaneous.image_c import post_image
+from .miscellaneous.factory_image_c import post_factory_image
