+++ backend/api/views/utils.py
@@ -1,10 +1,8 @@
 import random
-from datetime import datetime
 
 from django.conf import settings
 from django.contrib.gis.geos import Point
 from django.contrib.gis.measure import D
-from PIL import Image, ExifTags
 
 from ..models import Factory
 
@@ -38,28 +36,3 @@ def _get_client_ip(request):
         ip = request.META.get('REMOTE_ADDR')
     return ip
 
-
-def _get_image_original_date(f_image):
-    img = Image.open(f_image)
-
-    exif_raw = img._getexif()
-    if exif_raw is None:
-        return None
-
-    exif = {}
-    for k, v in exif_raw.items():
-        if k in ExifTags.TAGS:
-            exif[ExifTags.TAGS[k]] = v
-
-    try:
-        return datetime.strptime(exif["DateTimeOriginal"], "%Y:%m:%d %H:%M:%S")
-    except:
-        return None
-
-
-def _is_image(f_image):
-    try:
-        Image.open(f_image)
-        return True
-    except IOError:
-        return False
